requirements.txt
click==8.0.1
python-dotenv==0.19.1
openai==0.27.8
rich==13.4.1
